A few things in this commit: 1. Lobby and game events are completely separate now. Use /gameevent to set the game event. 2. The event tag in the ship config has changed. It now has two attributes, lobby and game, rather than the old single num attribute. 3. A new tag has been added to the ship config, namely <bans>. Its ignored at the moment.
Because of the changes to the ship config parsing, the DTD has been updated to v1.1. The new doctype is as follows:

<!DOCTYPE ships PUBLIC "-//Sylverant//DTD Ship Configuration 1.1//EN"
  "http://sylverant.net/dtd/ship_config1.1/ship_config.dtd">

The old DTD still will be up and available for the forseeable future, but the config parser will not parse it properly due to the changes in the <event> tag.