Adding in support for defining a calendar of events in advance.
New DTD for the ship_config.xml:
<!DOCTYPE ships PUBLIC "-//Sylverant//DTD Ship Configuration 1.5//EN"
  "http://sylverant.net/dtd/ship_config1.5/ship_config.dtd">

Note that this changes how events are defined, and if you use the ship_config v1.5, you must adopt the new event configuration (the old <event> tag is removed from it).
To keep things as they are, if you had a line like:
    <event game="0" lobby="0" />
in your old config, replace it with the following:
    <events>
        <defaults game="0" lobby="0" />
    </events>

Note that the parsing code still will handle the ship_config v1.4, so there's no need to upgrade right away.