Change around some stuff in the ship config...
This commit adds two new tags to the ship config, namely <bbmaps> and <bbparam>.
Also, this adds the ability to have both the shipgate address and the address of the ship itself looked up via DNS.

To use any of this new functionality, you'll need to update the DTD on the ship config XML file, as follows:
<!DOCTYPE ships PUBLIC "-//Sylverant//DTD Ship Configuration 1.8//EN"
  "http://sylverant.net/dtd/ship_config1.8/ship_config.dtd">

For the addresses, simply change any "ip" attributes in the shipgate and net tags to addr, and put in the FQDN as shown below:
    <shipgate addr="[shipgate FQDN]" port="[port]" ca="[ca cert]" />

Also, likewise for the <net> tag inside <ship>. Also, for the net tag, you can specify separate FQDNs for IPv4 and IPv6 by using addr and addr6 respectively. If you do not specify an addr6, it will try to use the addr for that as well.