In comes r600... Not much to see here for a "milestone" revision, but whatever... Here's what is included: 1. Fix a bug with word selects about meseta being detected as invalid. 2. Send blank C-Rank data about PSOv1 and Blue Burst clients to other versions so they don't get phantom C-Ranks. 3. Fix experience calculation for Blue Burst clients that don't kill an enemy, but do hit it. 4. Fix the enemy hit subcommand so it actually gets sent on Blue Burst. 5. Cleanups regarding bad memory use/not freeing stuff/etc. 6. Fix joining games on Blue Burst (so people can actually play together).
With this, Blue Burst actually becomes somewhat playable online with others. There's still a lot missing (like quests), but it is at least playable, in general. That said, its not anywhere near as well tested as the other versions at this point, and I'm sure there are lots of bugs still...