Make it so the DC NTE can at least get as far as the lobby.
This commit makes it so the DC NTE can at least get as far as the lobby without worrying about crashing due to other players being in the lobby when it joins or after the fact. This doesn't handle most things, and chatting is really b0rked with the NTE still. Progress is progress though.