Fix chatting on the DC NTE... mostly.
I'm not entirely 100% sure what should be after the '>' that I add to the string, if it even matters. I do know that whatever is after it never gets printed.

Note that there is still a problem with this if there is more than one character with the same name in the lobby. The jury is still out on whether that is fixable with regard to how the game is written or not.